@@ -2,22 +2,22 @@ package org.oxerr.seatgeek.client.rescu;
 
 import java.io.IOException;
 
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.PATCH;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.MediaType;
-
 import org.oxerr.seatgeek.model.request.CreateListingRequest;
 import org.oxerr.seatgeek.model.request.UpdateListingRequest;
 import org.oxerr.seatgeek.model.response.MultipleListingsResponse;
 import org.oxerr.seatgeek.model.response.Response;
 import org.oxerr.seatgeek.model.response.SingleListingResponse;
+
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.PATCH;
+import jakarta.ws.rs.PUT;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.QueryParam;
+import jakarta.ws.rs.core.MediaType;
 
 @Path("/")
 public interface ListingResource {
